Refactor TypicalAddressBook for test purposes
Currently, getTypicalAddressBook() only creates a example address book
with the TypicalMusicians. But we will want an address book with both
musicians and bands.

Hence, I refactored TypicalBands and TypicalMusicians into one
subpackage "typicalentities" in "/testutil". I have also refactored the
method getTypicalAddressBook() into a new class called
TypicalAddressBook.
