+++ src/main/java/seedu/address/logic/commands/FindBandCommand.java
@@ -33,9 +33,10 @@ public class FindBandCommand extends Command {
         requireNonNull(model);
         model.updateFilteredBandMusicianList(predicate);
 
-        // If the band exists, filtered band list is guaranteed to have only one band
+        // If the band exists, filtered band list is guaranteed to have only one band,
+        // (because add a band enforce no band with the same name (case-insensitive) is allowed).
         // If filtered band list size > 1 or size == 1 but the band filtered does not pass the predicate,
-        // it means that the band name is invalid
+        // it means that the band name is invalid, exception is thrown.
         if (model.getFilteredBandList().size() > 1 || !predicate.test(model.getFilteredBandList().get(0))) {
             throw new CommandException(Messages.MESSAGE_UNKNOWN_BAND);
         }
