+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -7,7 +7,7 @@ import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
 import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
 import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
 import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalMusicians.AMY;
+import static seedu.address.testutil.typicalentities.TypicalMusicians.AMY;
 
 import java.io.IOException;
 import java.nio.file.AccessDeniedException;
