+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalMusicians.ALICE;
+import static seedu.address.testutil.typicalentities.TypicalMusicians.ALICE;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
