+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -25,8 +25,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
-import static seedu.address.testutil.TypicalMusicians.AMY;
-import static seedu.address.testutil.TypicalMusicians.BOB;
+import static seedu.address.testutil.typicalentities.TypicalMusicians.AMY;
+import static seedu.address.testutil.typicalentities.TypicalMusicians.BOB;
 
 import org.junit.jupiter.api.Test;
