+++ src/test/java/seedu/address/storage/JsonAdaptedMusicianTest.java
@@ -3,7 +3,7 @@ package seedu.address.storage;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.address.storage.JsonAdaptedMusician.MISSING_FIELD_MESSAGE_FORMAT;
 import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalMusicians.BENSON;
+import static seedu.address.testutil.typicalentities.TypicalMusicians.BENSON;
 
 import java.util.ArrayList;
 import java.util.List;
